Move tree components to their own dir
