+++ extensions/vscode/webviews/homeView/src/components/tree/PostDecor.vue
