+++ extensions/vscode/webviews/homeView/src/components/tree/TreeItem.vue
