+++ extensions/vscode/webviews/homeView/src/components/tree/TreeSection.vue
